Fix EM doc string
