+++ docs/api/exposuremanagement/attack_path/index.rst
@@ -1,0 +1,1 @@
+.. automodule:: tenable.exposuremanagement.attack_path.api
