+++ docs/api/exposuremanagement/attack_path/vectors.rst
@@ -1,0 +1,1 @@
+.. automodule:: tenable.exposuremanagement.attack_path.vectors.api
