+++ tenable/exposuremanagement/attack_path/vectors/schema.py
