+++ tenable/exposuremanagement/inventory/schema.py
