+++ tenable/exposuremanagement/inventory/software/schema.py
